Using `uint64_t` instead of `int` and `int64_t` to avoid overflow

+++ runtime/util/finish_rewriting.cpp
@@ -15,7 +15,7 @@ bool statistics = false;
 bool binary_output = false;
 bool proof_output = false;
 
-int64_t get_steps();
+uint64_t get_steps();
 extern bool safe_partial;
 extern bool proof_hint_instrumentation_slow;
 
@@ -51,7 +51,7 @@ void init_outputs(char const *output_filename) {
   }
 
   if (statistics) {
-    int steps = get_steps();
+    uint64_t steps = get_steps();
     print_statistics(output_file, steps);
   }
